requirements.txt
flask==2.3.3
numpy==1.24.3
requests==2.31.0
python-telegram-bot==20.7
pandas==2.0.3
gunicorn==21.2.0